+++ spicerack/tests/unit/test_ganeti.py
@@ -2,26 +2,32 @@
 
 import json
 
+from unittest import mock
+
 import pytest
 import requests
 
 from spicerack.exceptions import SpicerackError
-from spicerack.ganeti import CLUSTER_SVC_URL, CLUSTERS_AND_ROWS, Ganeti, GanetiError, GanetiRAPI
-from spicerack.tests import get_fixture_path, requests_mock_not_available
+from spicerack.ganeti import CLUSTER_SVC_URL, CLUSTERS_AND_ROWS, Ganeti, GanetiError, GanetiRAPI, GntInstance
+from spicerack.remote import Remote
+from spicerack.tests import get_fixture_path, require_requests_mock
 
 
+@require_requests_mock
 class TestGaneti:
     """Ganeti tests class."""
 
     def setup_method(self):
         """Setup test environment."""
         # pylint: disable=attribute-defined-outside-init
-
-        self.ganeti = Ganeti(username='fake', password='password123', timeout=10)  # nosec
+        self.remote = mock.MagicMock(spec_set=Remote)
+        self.ganeti = Ganeti(username='user', password='pass', timeout=10, remote=self.remote)  # nosec
 
         self.cluster = 'eqiad'
+        self.instance = 'test.example.com'
         self.cluster_base_url = {cluster: CLUSTER_SVC_URL.format(dc=cluster) + '/2' for cluster in CLUSTERS_AND_ROWS}
         self.base_url = self.cluster_base_url[self.cluster]
+        self.instance_url = '{base_url}/instances/{instance}'.format(base_url=self.base_url, instance=self.instance)
 
         # load test fixtures
         with open(get_fixture_path('ganeti', 'info.json'), encoding='utf-8') as info_json:
@@ -33,18 +39,27 @@ class TestGaneti:
         with open(get_fixture_path('ganeti', 'bogus.json'), encoding='utf-8') as bogus_json:
             self.bogus_data = bogus_json.read()
 
+    def _set_requests_mock_for_instance(self, requests_mock, missing_active=False):
+        """Set request mock to be 404 on all other clusters."""
+        for cluster in CLUSTERS_AND_ROWS:
+            if cluster == self.cluster and not missing_active:
+                requests_mock.get(self.instance_url, text=self.instance_info)
+            else:
+                requests_mock.get('{base_url}/instances/{instance}'.format(
+                    base_url=self.cluster_base_url[cluster], instance=self.instance),
+                    text=self.fourohfour, status_code=requests.codes['not_found'])
+
     @pytest.mark.parametrize('cluster', CLUSTERS_AND_ROWS.keys())
-    def test_ganeti_clusters(self, cluster):
+    def test_rapi_clusters_ok(self, cluster):
         """Ganeti.rapi() should return a GanetiRAPI object."""
         assert isinstance(self.ganeti.rapi(cluster), GanetiRAPI)
 
-    def test_ganeti_clusters_failure(self):
+    def test_rapi_clusters_missing(self):
         """If a cluster that doesn't exist is passed to Ganeti.rapi(), it should except with a SpicerackError."""
         with pytest.raises(SpicerackError, match=r'Cannot find cluster bogus cluster \(expected .*'):
             self.ganeti.rapi('bogus cluster')
 
-    @pytest.mark.skipif(requests_mock_not_available(), reason='Requires requests-mock fixture')
-    def test_ganeti_rapi_master_success(self, requests_mock):
+    def test_rapi_master_ok(self, requests_mock):
         """The master property of a RAPI should be the hostname for the master of this cluster."""
         requests_mock.get(self.base_url + '/info', text=self.info)
 
@@ -52,67 +67,86 @@ class TestGaneti:
         rapi = self.ganeti.rapi(self.cluster)
         assert rapi.master == master
 
-    @pytest.mark.skipif(requests_mock_not_available(), reason='Requires requests-mock fixture')
-    def test_ganeti_rapi_master_failure(self, requests_mock):
+    def test_rapi_master_missing(self, requests_mock):
         """If a master is not specified by the uptsream API, the value of master on a RAPI should be None."""
         requests_mock.get(self.base_url + '/info', text=self.bogus_data)
         rapi = self.ganeti.rapi(self.cluster)
         assert rapi.master is None
 
-    @pytest.mark.skipif(requests_mock_not_available(), reason='Requires requests-mock fixture')
-    def test_ganeti_rapi_instance_notfound(self, requests_mock):
+    def test_rapi_instance_missing(self, requests_mock):
         """A RAPI object should raise a GanetiError if a requested host does not exist."""
+        requests_mock.get(self.instance_url, text=self.fourohfour, status_code=requests.codes['not_found'])
         rapi = self.ganeti.rapi(self.cluster)
-        requests_mock.get(
-            self.base_url + '/instances/testhost_404', text=self.fourohfour, status_code=requests.codes['not_found']
-        )
         with pytest.raises(GanetiError, match=r'Non-200 from API: 404:.*'):
-            rapi.fetch_instance('testhost_404')
+            rapi.fetch_instance(self.instance)
 
-    @pytest.mark.skipif(requests_mock_not_available(), reason='Requires requests-mock fixture')
-    def test_ganeti_rapi_instance_timeout(self, requests_mock):
+    def test_rapi_instance_timeout(self, requests_mock):
         """A RAPI object should raise a GanetiError if a request times out."""
         rapi = self.ganeti.rapi(self.cluster)
-        requests_mock.get(self.base_url + '/instances/timeouthost', exc=requests.exceptions.ConnectTimeout)
+        requests_mock.get(self.instance_url, exc=requests.exceptions.ConnectTimeout)
         with pytest.raises(GanetiError, match='Timeout performing request to RAPI'):
-            rapi.fetch_instance('timeouthost')
+            rapi.fetch_instance(self.instance)
 
-    @pytest.mark.skipif(requests_mock_not_available(), reason='Requires requests-mock fixture')
-    def test_ganeti_rapi_instance_invalid(self, requests_mock):
+    def test_rapi_instance_invalid(self, requests_mock):
         """If no mac is present in host data, fetch_instance_mac should raise a GanetiError."""
         rapi = self.ganeti.rapi(self.cluster)
-        requests_mock.get(self.base_url + '/instances/testhost_invalid', text=self.bogus_data)
+        requests_mock.get(self.instance_url, text=self.bogus_data)
         with pytest.raises(GanetiError, match=''):
-            rapi.fetch_instance_mac('testhost_invalid')
+            rapi.fetch_instance_mac(self.instance)
 
-    @pytest.mark.skipif(requests_mock_not_available(), reason='Requires requests-mock fixture')
-    def test_ganeti_rapi_instance_valid(self, requests_mock):
+    def test_rapi_instance_valid(self, requests_mock):
         """The MAC returned by RAPI.fetch_instance_mac should match the data returned by the API."""
         rapi = self.ganeti.rapi(self.cluster)
-        requests_mock.get(self.base_url + '/instances/testhost', text=self.instance_info)
+        requests_mock.get(self.instance_url, text=self.instance_info)
         mac = json.loads(self.instance_info)['nic.macs'][0]
-        assert mac == rapi.fetch_instance_mac('testhost')
+        assert mac == rapi.fetch_instance_mac(self.instance)
 
-    @pytest.mark.skipif(requests_mock_not_available(), reason='Requires requests-mock fixture')
-    def test_ganeti_get_cluster_for_instance(self, requests_mock):
+    def test_fetch_cluster_for_instance_ok(self, requests_mock):
         """We should get a cluster name that the host exists for."""
-        requests_mock.get(self.base_url + '/instances/testhost', text=self.instance_info)
-        for cluster in CLUSTERS_AND_ROWS:
-            if cluster == self.cluster:
-                continue
-            requests_mock.get(
-                self.cluster_base_url[cluster] + '/instances/testhost', text=self.fourohfour,
-                status_code=requests.codes['not_found']
-            )
-        assert self.ganeti.fetch_cluster_for_instance('testhost') == self.cluster
+        self._set_requests_mock_for_instance(requests_mock)
+        assert self.ganeti.fetch_cluster_for_instance(self.instance) == self.cluster
 
-    @pytest.mark.skipif(requests_mock_not_available(), reason='Requires requests-mock fixture')
-    def test_ganeti_get_cluster_for_instance_notfound(self, requests_mock):
+    def test_fetch_cluster_for_instance_missing(self, requests_mock):
         """We should get an exception if the host is not found in Ganeti."""
-        for cluster in CLUSTERS_AND_ROWS:
-            requests_mock.get(
-                self.cluster_base_url[cluster] + '/instances/testhost', text=self.fourohfour,
-                status_code=requests.codes['not_found']
-            )
-        with pytest.raises(GanetiError, match='Cannot find testhost in any configured cluster.'):
-            self.ganeti.fetch_cluster_for_instance('testhost')
+        self._set_requests_mock_for_instance(requests_mock, missing_active=True)
+        with pytest.raises(GanetiError, match='Cannot find test.example.com in any configured cluster.'):
+            self.ganeti.fetch_cluster_for_instance(self.instance)
+
+    @pytest.mark.parametrize('cluster', ('', 'eqiad'))
+    def test_instance_ok(self, cluster, requests_mock):
+        """It should return an instance of GntInstance for a properly configured cluster."""
+        self._set_requests_mock_for_instance(requests_mock)
+        requests_mock.get(self.base_url + '/info', text=self.info)
+        instance = self.ganeti.instance(self.instance, cluster=cluster)
+        assert isinstance(instance, GntInstance)
+        assert instance.cluster == 'eqiad'
+        self.remote.query.assert_called_once_with('ganeti1.example.com')
+
+    def test_instance_missing_master(self, requests_mock):
+        """It should raise a GanetiError exception if unable to determin the instance's master to manage it."""
+        self._set_requests_mock_for_instance(requests_mock)
+        requests_mock.get(self.base_url + '/info', text=self.bogus_data)
+        with pytest.raises(GanetiError, match='Master for cluster eqiad is None'):
+            self.ganeti.instance(self.instance)
+
+    @pytest.mark.parametrize('kwargs', ({}, {'timeout': 0}))
+    def test_instance_shutdown(self, requests_mock, kwargs):
+        """It should issue the shutdown command on the master host."""
+        self._set_requests_mock_for_instance(requests_mock)
+        requests_mock.get(self.base_url + '/info', text=self.info)
+        instance = self.ganeti.instance(self.instance)
+        instance.shutdown(**kwargs)
+        timeout = kwargs['timeout'] if 'timeout' in kwargs else 2
+        self.remote.query.return_value.run_sync.assert_called_once_with(
+            'gnt-instance shutdown --timeout={timeout} test.example.com'.format(timeout=timeout))
+
+    @pytest.mark.parametrize('kwargs', ({}, {'shutdown_timeout': 0}))
+    def test_instance_remove(self, requests_mock, kwargs):
+        """It should issue the remove command on the master host."""
+        self._set_requests_mock_for_instance(requests_mock)
+        requests_mock.get(self.base_url + '/info', text=self.info)
+        instance = self.ganeti.instance(self.instance)
+        instance.remove(**kwargs)
+        timeout = kwargs['shutdown_timeout'] if 'shutdown_timeout' in kwargs else 2
+        self.remote.query.return_value.run_sync.assert_called_once_with(
+            'gnt-instance remove --shutdown-timeout={timeout} --force test.example.com'.format(timeout=timeout))